Adds dateRange prop to MediaLibrary component trees and query
Previously the selected date range was not communicated outside of the MediaDateRange component. Adds a dateRange prop and onDateChange callback prop which are used to store the current date range and update the state of the dateRange prop respectively. This enables us to pass the dateRange into the API query used to retrieve photos data.